refactor: moved interfaces and types to model file
