+++ src/public_api.ts
@@ -2,3 +2,4 @@ export * from './lib/actions';
 export * from './lib/entity-state';
 export * from './lib/errors';
 export * from './lib/id-strategy';
+export * from './lib/models';
